Clean up ownership
- Don't loan widgets (NavBar, SystemBar, Activities)
- Prefer to pass dependencies into constructor
- Clean up ownership in Widget hierarchy
- Finish onAttached / onDetached support in Widgets
- More unique_ptr

+++ ocher/fmt/text/Text.cpp
@@ -16,5 +16,5 @@ Text::Text(const std::string& filename)
     File f(filename);
 
     f.readRest(m_text);
-    Log::debug(LOG_NAME, "Loaded %lu bytes", (unsigned long)m_text.size());
+    Log::debug(LOG_NAME, "Loaded %zu bytes", m_text.size());
 }
